decode: add ns_keyed_archiver, restructure apple decoder into apple package
