+++ format/all/all.go
@@ -5,7 +5,8 @@ package all
 
 import (
 	_ "github.com/wader/fq/format/ape"
-	_ "github.com/wader/fq/format/applebookmark"
+	_ "github.com/wader/fq/format/apple/bookmark"
+	_ "github.com/wader/fq/format/apple/bplist"
 	_ "github.com/wader/fq/format/ar"
 	_ "github.com/wader/fq/format/asn1"
 	_ "github.com/wader/fq/format/av1"
@@ -13,7 +14,6 @@ import (
 	_ "github.com/wader/fq/format/bencode"
 	_ "github.com/wader/fq/format/bitcoin"
 	_ "github.com/wader/fq/format/bits"
-	_ "github.com/wader/fq/format/bplist"
 	_ "github.com/wader/fq/format/bson"
 	_ "github.com/wader/fq/format/bzip2"
 	_ "github.com/wader/fq/format/cbor"
