+++ format/apple/bookmark/apple_bookmark.go
@@ -1,4 +1,4 @@
-package applebookmark
+package bookmarkdata
 
 import (
 	"embed"
